requirements.txt
customtkinter>=5.2.0
selenium>=4.15.0
python-chess>=1.999
stockfish>=3.28.0
webdriver-manager>=4.0.0